Make the service worker ready to handle WordPress requests immediately after the registration.
This commit makes the service worker claim its client upon the
`activate` event. This is necessary to set it as the "controller"
of the page that requested the registration. Without the claim() call
the service worker only becomes the "controller" after a page reload,
meaning it isn't used to resolve requests and the initial WordPres
render shows a 404 error.