Throw an exception when an infinite loading loop is detected.
When the service worker fails for any reason, the page displayed inside the iframe won't be a WordPress instance we expect from the service worker. Instead, it will be the original page trying to load the service worker. This causes an infinite loop with a loader inside a loader inside a loader.