Handle „/„ requests as PHP requests in the service worker
Before this commit, the service worker would ignore requests to URLs like `/` or `/scope:90841`. This behavior is a legacy artifact from the time when the main HTML file was called index.html and we wanted to avoid conflicts between requests intended for `index.html` and WordPress’s index.php. Since the `index.html` was since renamed to `wordpress.html` specifically to avoid these conflicts, we can now safely treat all `/` requests as intended for WordPress.