Make the PHP class constructor private
The streams argument is so implementeation-specific that the larger PHP class doesn’t make much sense without it. Let’s hide it until a more general API emerges.

This is an attempt to avoid having an asynchronous init() method that must be called before the PHP class can be used. It removes an entire class of errors related to calling different methods before that async init() finishes.