PHP: Output stdout and stderr to files instead of streams
Replaces output streams with temporary files to fix the following test case:

```
  ● PHP  › should echo bytes
    expect(
        php.run('<?php echo chr(1).chr(0).chr(1).chr(0).chr(2);').stdout
    ).toBe('\x01\x00\x01\x00\x02');

    Expected: "1 0 1 0 2"
    Received: "1 1 2"
```

Turns out Emscripten has special semantics for null bytes in stdout and doesn't pass them to JavaScript handlers. This commit ditches stdout and stderr as a transport, and moves to  temporary files instead. This way the null bytes output by PHP make it unchanged to the browser.

One other change to enable that is upgrading the PHPOutput.stdout datatype from string to ArrayBuffer to make sure the bytes are passed around correctly.

Solves #80