Replace BroadcastChannel with regular postMessage
BroadcastChannel breaks iframe embedding the playground in Safari.

Weirdly, Safari does not pass any messages from the ServiceWorker to Window if the page is rendered inside an iframe. Window to Service Worker communication works just fine.

The regular client.postMessage() communication works perfectly, so this commit uses that instead of the BroadcastChannel.