Asyncify: Only add lastAsyncifyStackSource callback if Asyncify is defined (#399)
Web version of PHP.wasm is built without Asyncify. However, attaching
the lastAsyncifyStackSource assumes Asyncify is defined and the code
errors out. This commit adds a safety check to the
append-before-return.js file. It was already added to each PHP
JavaScript module, but somehow didn't make it to the source file from
which these modules are generated.