Add deferred loading mode with "Run Playground" button (#537)
By default the Playground website starts downloading all of the Playground
assets, including React, PHP compiled to WASM, and the WordPress PHP files.

With this change there's a new `?lazy` query argument that if set will
present a "Run Playground" button instead of the actual Playground. When
clicked, this button will trigger the full load for the Playground.

Co-authored-by: Adam Zielinski <adam@adamziel.com>
Co-authored-by: Dennis Snell <dennis.snell@automattic.com>