[Web] Don't force-reload the Service Worker (#561)
Removes the entire concept of detecting the service worker version and enforcing an update on mismatch.

Auto-reloading the service worker when it's updated thrashes any `postMessage` communication that's in progress at the moment of calling `skipWaiting()`. This causes Playground to hang at the "login" step in https://github.com/WordPress/wordpress-playground/pull/559.

However, the browsers handle a lot by default:

* `registration.update()` method downloads the new service-worker.js file and compares it byte-by-byte with the existing one
* The previous service worker won't die until all the browser tabs it serves are closed
* The new service worker will automatically replace the previous one afterwards

The only problem remains deploying a website that is backwards–incompatible with the previous service worker. This is tracked separately in https://github.com/WordPress/wordpress-playground/issues/566

Co-authored-by: Dennis Snell <dennis.snell@automattic.com>