Fake websocket connection on web to prevent creating WebSocket for SOCKFS
When initializing the `SOCKFS` from emcripten, a websocket is created automatically
whether it's needed or not (whether any filesystems are `mount()`ed into the `SOCKFS`).

This can, from time to time, create errors on initialization which _might_ block
loading of the Playground. These errors are benign because the websocket isn't needed
or supported, but they are _thrown_ and so still cause problems.

In this patch we're providing a fake WebSocket class through the decorator mechanism
so that any errors in the constructor are swallowed and made safe.